+++ utils.py
@@ -1,0 +1,6 @@
+from json import JSONEncoder
+
+
+class StandardEncoder(JSONEncoder):
+    def default(self, o):
+        return o.to_dict()
